requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
statsmodels>=0.13.0
tensorflow>=2.8.0
seaborn>=0.11.0
jupyter>=1.0.0
pyyaml>=5.4.0
scipy>=1.7.0
